tbrekalo-chain: updating nn eval
