+++ notebooks/nn.py
@@ -1,0 +1,3 @@
+"""Module with neural network utilities."""
+
+import torch
